comment on possible error and it's correction
The original document claims that for erase display and erase line, the default parameter is 1 and the direction erased is for 0 and 1 the reverse  of what it is for Linux.  In Linux, 0 is the default parameter and applied to what was written here, 1 is 0, and 0 is 1.  Hopefully this is a mistake , but I added comentation on it and need to test the problem on all of the graphical VT100 or VT220 supporting consoles.